chore: refactor/remove mixin
